requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.26.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
